requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
cloudscraper
aiohttp
openai>=1.0.0